renamed files to index overlay
