+++ frontend/src/components/graph2D/indexOverlay/IndexOverlay.module.scss
