+++ frontend/src/components/graph2D/indexOverlay/IndexOverlay.tsx
